requirements.txt
pyautogui
Pillow
pyperclip
keyboard
copykitten